+++ tests/_proxies/Assets/Manager.php
@@ -132,4 +132,9 @@ class Manager extends PhManager
     {
         return parent::outputInlineJs($collectionName);
     }
+
+    public function exists($id)
+    {
+        return parent::exists($id);
+    }
 }