+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - Added `Phalcon\Validation\Validator\Date`
 - Mcrypt is replaced with openssl in `Phalcon\Crypt`
 - Removed methods setMode(), getMode(), getAvailableModes() in `Phalcon\CryptInterface`
+- Added `Phalcon\Assets\Manager::exists()` to check if collection exists
 
 # [2.0.11](https://github.com/phalcon/cphalcon/releases/tag/phalcon-v2.0.11) (????-??-??)
 - Fix Model magic set functionality to maintain variable visibility and utilize setter methods.[#11286](https://github.com/phalcon/cphalcon/issues/11286)